feat: support cleanDist option

+++ src/make.ts
@@ -7,6 +7,7 @@ interface mkdistOptions {
   rootDir?: string
   srcDir?: string
   distDir?: string
+  cleanDist?: boolean
   format?: CreateLoaderOptions['format']
 }
 
@@ -17,9 +18,11 @@ export async function mkdist (options: mkdistOptions /* istanbul ignore next */ 
   options.distDir = resolve(options.rootDir, options.distDir || 'dist')
 
   // Setup dist
-  await unlink(options.distDir).catch(() => {})
-  await emptyDir(options.distDir)
-  await mkdirp(options.distDir)
+  if (options.cleanDist !== false) {
+    await unlink(options.distDir).catch(() => {})
+    await emptyDir(options.distDir)
+    await mkdirp(options.distDir)
+  }
 
   const filePaths = await globby('**', { absolute: false, cwd: options.srcDir })
